extract DownloadsListComponent
