+++ app/components/work_oh_audio_show_component.html.erb
@@ -13,11 +13,11 @@
         <h1 class="title h6">
           <%= work.title %>
         </h1>
-        <% if combined_mp3_audio.present? && derivatives_up_to_date? %>
+        <% if combined_mp3_audio_url.present? && combined_derivatives_up_to_date? %>
           <div class="d-flex" style="align-items: center">
             <audio controls controlsList="nodownload" data-role="ohms-audio-elem" >
-                <source src="<%= combined_mp3_audio%>"  type="audio/mpeg"/>
-                <source src="<%= combined_webm_audio%>"  type="audio/webm"/>
+                <source src="<%= combined_mp3_audio_url%>"  type="audio/mpeg"/>
+                <source src="<%= combined_webm_audio_url%>"  type="audio/webm"/>
             </audio>
             <% if has_ohms_index? || has_ohms_transcript? %>
               <div class="pl-3">
@@ -172,7 +172,7 @@
         <% end %>
 
         <div class="tab-pane downloads" id="ohDownloads" role="tabpanel" aria-labelledby="ohDownloadsTab">
-          <%= render 'show_with_audio_downloads', decorator: self %>
+          <%= render OralHistory::DownloadsListComponent.new(work: work, decorator: self) %>
         </div>
       </div>
   </div>
